refactor: split formatter.rs into submodules
